move common test helper routines into testutil.cpp
